+++ extensions/elasticsearch/ActiveQuery.php
@@ -64,10 +64,12 @@ use yii\db\ActiveRelationTrait;
  * This methods may only be called in a relational context. Same is true for [[inverseOf()]], which
  * marks a relation as inverse of another relation.
  *
- * > NOTE: elasticsearch limits the number of records returned by any query to 10 records by default.
+ * > Note: elasticsearch limits the number of records returned by any query to 10 records by default.
  * > If you expect to get more records you should specify limit explicitly in relation definition.
  * > This is also important for relations that use [[via()]] so that if via records are limited to 10
  * > the relations records can also not be more than 10.
+ *
+ * > Note: Currently [[with]] is not supported in combination with [[asArray]].
  *
  * @author Carsten Brandt <mail@cebe.cc>
  * @since 2.0